refactor(WalletConnect): screens
