+++ src/debug-vars.ts
@@ -1,3 +1,4 @@
 export const DEBUG_VARS = {
   enableSentry: !__DEV__,
+  enableWalletConnectLogger: false,
 };
