+++ src/screens/wallet-connect-sign.tsx
@@ -1,23 +1,14 @@
 import React, {useCallback, useMemo} from 'react';
 
-import {ScrollView, View} from 'react-native';
-import {useSafeAreaInsets} from 'react-native-safe-area-context';
-
-import {Color} from '@app/colors';
-import {Button, ButtonVariant, Spacer} from '@app/components/ui';
-import {WalletConnectSignInfo} from '@app/components/wallet-connect-sign-info';
-import {WalletConnectTransactionInfo} from '@app/components/wallet-connect-transaction-info';
-import {createTheme} from '@app/helpers';
+import {WalletConnectSign} from '@app/components/wallet-connect-sign';
 import {useTypedNavigation, useTypedRoute} from '@app/hooks';
 import {useWalletConnectSession} from '@app/hooks/use-wallet-connect-session';
-import {I18N} from '@app/i18n';
 import {Wallet} from '@app/models/wallet';
 import {WalletConnect} from '@app/services/wallet-connect';
 import {getUserAddressFromSessionRequest} from '@app/utils';
 import {EIP155_SIGNING_METHODS} from '@app/variables/EIP155';
 
 export const WalletConnectSignScreen = () => {
-  const insets = useSafeAreaInsets();
   const navigation = useTypedNavigation();
   const route = useTypedRoute<'walletConnectSign'>();
   const event = useMemo(() => route?.params?.event, [route?.params?.event]);
@@ -55,58 +46,13 @@ export const WalletConnectSignScreen = () => {
   }, [event, navigation]);
 
   return (
-    <View style={styles.container}>
-      <ScrollView
-        bounces={false}
-        showsVerticalScrollIndicator={false}
-        style={styles.scroll}
-        contentContainerStyle={styles.scrollContainer}>
-        <Spacer height={32} />
-
-        {isTransaction && (
-          <WalletConnectTransactionInfo event={event} session={session!} />
-        )}
-
-        {!isTransaction && (
-          <WalletConnectSignInfo
-            event={event}
-            session={session!}
-            wallet={wallet!}
-          />
-        )}
-      </ScrollView>
-
-      <View style={[styles.buttonContainer, {marginBottom: insets.bottom}]}>
-        <Spacer height={4} />
-        <Button
-          variant={ButtonVariant.contained}
-          onPress={onPressSign}
-          i18n={I18N.walletConnectSignApproveButton}
-        />
-        <Button
-          textColor={Color.textRed1}
-          onPress={onPressReject}
-          i18n={I18N.walletConnectSignRejectButton}
-        />
-      </View>
-    </View>
+    <WalletConnectSign
+      isTransaction={isTransaction}
+      event={event}
+      session={session!}
+      wallet={wallet!}
+      onPressReject={onPressReject}
+      onPressSign={onPressSign}
+    />
   );
 };
-
-const styles = createTheme({
-  container: {
-    paddingHorizontal: 20,
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    flex: 1,
-  },
-  scrollContainer: {
-    alignItems: 'center',
-  },
-  scroll: {
-    width: '100%',
-  },
-  buttonContainer: {
-    width: '100%',
-  },
-});
